zsh: Kick syntax highlighting after restoring previous ZLE buffer
When navigate-gitroot-or-home ZLE widget causes a send-break, it defers buffer
restoration until the next zle-line-init. Zsh fast-syntax-highlighting
incorrectly does not invalidate its buffer memory on a full break/init cycle,
so when we restore an identical buffer later, it does not try to highlight it.

Implements a simple function which nukes its buffer cache and then invokes its
highlighting function again. We call this any time we're restoring the buffer
via the restore-buffer plugin. Might be useful in other places too.

This would be a much simpler fix upstream to f-sy-h, but its developer has been
MIA for ages, and various public forks are in varying states of
non-maintenance. I'd rather not carry a diff on our vendored copy.