bin/terminal: Launch as floating by default
Invert our own call sites to manually request a tiled terminal.

This is useful because nearly all of the time when a terminal is launched by
another application through the XDG/mimeapps.list mechanism (such as Vivaldi/
Chromium showing the containing folder of a download), what we prefer is a
floating terminal.