skhd: Add binding for 60% split
Existing binding to invert 70/30 split direction wasn't very useful. On the
other hand, some apps are not quite usable in the smaller portion of a 70/30
split, but we don't want to have to go all the way for a 50/50 and waste space.
This way, the keys []\ map to 50/60/70 split respectively (assuming two main
windows, since \ is actually a horizontal balance).